fix(container): repair the cron settings
When building the container, the env var containing the cron settings was messed up and the app to never re-executed the ipChecker function.
EX: when -e CRON='*/15 * * * *' was set, only '*/15' was passed to the cli inside the container.
In order to prevent this, I split the cron config in several env vars